Moved AgendaCategories to Settings & removed Delete for action categories
