+++ public/locales/fr/translation.json
@@ -610,10 +610,11 @@
     "actionItemCategories": "Catégories d'éléments d'action",
     "updateOrganization": "Mettre à jour l'organisation",
     "seeRequest": "Voir la demande",
-    "noData": "Pas de données",
-    "otherSettings": "Autres réglages",
+    "noData": "Aucune donnée",
+    "otherSettings": "Autres paramètres",
     "changeLanguage": "Changer de langue",
-    "manageCustomFields": "Gérer les champs personnalisés"
+    "manageCustomFields": "Gérer les champs personnalisés",
+    "agendaItemCategories": "Catégories d'éléments d'agenda"
   },
   "deleteOrg": {
     "deleteOrganization": "Supprimer l'organisation",
