+++ public/locales/hi/translation.json
@@ -605,15 +605,16 @@
     "amount": "मात्रा"
   },
   "orgSettings": {
-    "title": "समायोजन",
+    "title": "सेटिंग्स",
     "general": "सामान्य",
-    "actionItemCategories": "कार्रवाई आइटम श्रेणियाँ",
-    "updateOrganization": "संगठन अद्यतन करें",
+    "actionItemCategories": "कार्य आइटम श्रेणियाँ",
+    "updateOrganization": "संगठन अपडेट करें",
     "seeRequest": "अनुरोध देखें",
     "noData": "कोई डेटा नहीं",
-    "otherSettings": "अन्य सेटिंग",
+    "otherSettings": "अन्य सेटिंग्स",
     "changeLanguage": "भाषा बदलें",
-    "manageCustomFields": "कस्टम फ़ील्ड प्रबंधित करें"
+    "manageCustomFields": "कस्टम फ़ील्ड प्रबंधित करें",
+    "agendaItemCategories": "एजेंडा आइटम श्रेणियाँ"
   },
   "deleteOrg": {
     "deleteOrganization": "संगठन हटाएँ",
