+++ src/App.tsx
@@ -14,7 +14,6 @@ import OrgList from 'screens/OrgList/OrgList';
 import OrgPost from 'screens/OrgPost/OrgPost';
 import OrgSettings from 'screens/OrgSettings/OrgSettings';
 import OrganizationActionItems from 'screens/OrganizationActionItems/OrganizationActionItems';
-import OrganizationAgendaCategory from 'screens/OrganizationAgendaCategory/OrganizationAgendaCategory';
 import OrganizationDashboard from 'screens/OrganizationDashboard/OrganizationDashboard';
 import OrganizationEvents from 'screens/OrganizationEvents/OrganizationEvents';
 import OrganizaitionFundCampiagn from 'screens/OrganizationFundCampaign/OrganizationFundCampagins';
@@ -37,7 +36,6 @@ import Posts from 'screens/UserPortal/Posts/Posts';
 import Organizations from 'screens/UserPortal/Organizations/Organizations';
 import People from 'screens/UserPortal/People/People';
 import Settings from 'screens/UserPortal/Settings/Settings';
-// import UserLoginPage from 'screens/UserPortal/UserLoginPage/UserLoginPage';
 import Chat from 'screens/UserPortal/Chat/Chat';
 import { useQuery } from '@apollo/client';
 import { CHECK_AUTH } from 'GraphQl/Queries/Queries';
@@ -163,10 +161,6 @@ function app(): JSX.Element {
             <Route
               path="/orgactionitems/:orgId"
               element={<OrganizationActionItems />}
-            />
-            <Route
-              path="/orgagendacategory/:orgId"
-              element={<OrganizationAgendaCategory />}
             />
             <Route path="/orgfunds/:orgId" element={<OrganizationFunds />} />
             <Route
