+++ src/components/IconComponent/IconComponent.tsx
@@ -4,7 +4,6 @@ import {
   NewspaperOutlined,
 } from '@mui/icons-material';
 import { ReactComponent as ActionItemIcon } from 'assets/svgs/actionItem.svg';
-import { ReactComponent as AgendaCategoryIcon } from 'assets/svgs/agenda-category-icon.svg';
 import { ReactComponent as BlockUserIcon } from 'assets/svgs/blockUser.svg';
 import { ReactComponent as CheckInRegistrantsIcon } from 'assets/svgs/checkInRegistrants.svg';
 import { ReactComponent as DashboardIcon } from 'assets/svgs/dashboard.svg';
@@ -66,13 +65,6 @@ const iconComponent = (props: InterfaceIconComponent): JSX.Element => {
         <ActionItemIcon
           {...props}
           data-testid="Icon-Component-ActionItemIcon"
-        />
-      );
-    case 'Agenda Items Category':
-      return (
-        <AgendaCategoryIcon
-          {...props}
-          data-testid="Icon-Component-AgendaCategoryIcon"
         />
       );
     case 'Posts':
