+++ src/state/reducers/routesReducer.ts
@@ -77,11 +77,6 @@ const components: ComponentType[] = [
     comp_id: 'orgactionitems',
     component: 'OrganizationActionItems',
   },
-  {
-    name: 'Agenda Items Category',
-    comp_id: 'orgagendacategory',
-    component: 'OrganizationAgendaCategory',
-  },
   { name: 'Posts', comp_id: 'orgpost', component: 'OrgPost' },
   { name: 'Block/Unblock', comp_id: 'blockuser', component: 'BlockUser' },
   { name: 'Advertisement', comp_id: 'orgads', component: 'Advertisements' },
